requirements.txt
crewai
langchain
python-dotenv
openai
flask
pandas
statsmodels
gunicorn